Move suite commands to dir
