+++ setup.cfg
@@ -62,7 +62,7 @@ all =
 [options.entry_points]
 console_scripts =
     duct = con_duct.__main__:main
-    con-duct = con_duct.suite:main
+    con-duct = con_duct.suite.main:main
 
 [mypy]
 allow_incomplete_defs = False
